test(pubmed): fix extract_article_info
